python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
requests==2.31.0
